32-bit dates so you can go over 200 years in-game

+++ src/common_types/date_interface.hpp
@@ -55,7 +55,7 @@ public:
 
 class date {
 public:
-	using value_base_t = uint16_t;
+	using value_base_t = uint32_t;
 	using zero_is_null_t = std::true_type;
 
 	value_base_t value = 0;
